requirements.txt
fastapi
uvicorn[standard]
python-dotenv
line-bot-sdk
playwright==1.48.0
